test(dashboard): the flag-off subject boots on its own fixture page (#17948)
